refactor(models): split index.ts into lib modules
Pure relocation of the 897-line models plugin entry into cohesive lib
modules, mirroring the tasks/schedule thin-index template:

- lib/model-id.ts — pure id helpers (normalize/tier/provider); leaf module
  so config-io and available-models stay acyclic
- lib/config-io.ts — runtime config read/write/update, RuntimeModelConfig
  types, team-hook agent-meta cache, resolveAgents, and the restart-sync
  cell (markConfigDirty/markRuntimeRestarted — one owning module)
- lib/available-models.ts — in-memory models cache cell + disk-cache
  hydration + live fetch with in-flight dedupe + known-models catalog merge
- lib/aliases.ts — DEFAULT_ALIASES + readAliases
- lib/route-schemas.ts — zod request/response schemas + spend-window parse
- lib/routes.ts — the 14 declarative routes as modelsRoutes
- lib/register-hooks.ts — the nine models.* cross-plugin hooks
- lib/exec-tools.ts — the two read-only MCP exec tools

index.ts is now the thin definePlugin shell (44 lines). Bodies moved
byte-for-byte; hook/tool names, route paths, and the persisted cache path
are unchanged.
